lint warnings reduction, more import refactors
