+++ packages/bots/fuse-liquidator-bot/index.ts
@@ -1,5 +1,5 @@
 import { JsonRpcProvider } from "@ethersproject/providers";
-import { SupportedChains } from "@midas-capital/sdk";
+import { SupportedChains } from "@midas-capital/types";
 import dotenv from "dotenv";
 
 import { approveTokensToSafeLiquidator, liquidateAndRepeat, logger } from "./src";
