+++ packages/sdk/chainDeploy/helpers/chainlink.ts
@@ -1,7 +1,7 @@
+import { SupportedAsset } from "@midas-capital/types";
 import { providers } from "ethers";
 
 import { AddressesProvider } from "../../lib/contracts/typechain/AddressesProvider";
-import { SupportedAsset } from "../../src/types";
 
 import { ChainlinkDeployFnParams, ChainlinkFeedBaseCurrency } from "./types";
