+++ packages/sdk/chainDeploy/testnets/moonbase.ts
@@ -1,10 +1,9 @@
 /* eslint-disable no-console, @typescript-eslint/no-non-null-assertion */
 
+import { SupportedAsset, SupportedChains } from "@midas-capital/types";
 import { ethers, providers } from "ethers";
 
-import { SupportedChains } from "../../src";
 import { assetSymbols, chainSpecificParams, chainSupportedAssets } from "../../src/chainConfig";
-import { SupportedAsset } from "../../src/types";
 import { ChainDeployConfig } from "../helpers";
 import { ChainDeployFnParams } from "../helpers/types";
